requirements.txt



python-dotenv

pandas

google-cloud-bigquery

tweepy
